refactor: centralize model checkpointing utilities
